Implemented Printer
Utilized 4 maps that provided different functionality. The kindIndex simply gave the index of the
column which the print statement is called for. The other 3 maps simply store the information, utilizing
1 or 2 pairs inside the map. I wasn't sure if if the state was Finished we first print out the Finished
only with the "..." and then print out all the other stuff or print it at once. Right now im using the latter
as the assumption.